backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
Pillow==10.4.0
openai==1.35.0
anthropic==0.31.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
